Reorganise init helper function code chunk
